updated sortoptions dropdown to current designs
