+++ client/src/components/Sortoptions/Category.js
@@ -1,4 +1,6 @@
+
 import React from 'react';
+
 import { Typography } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
 import Menu from '@material-ui/core/Menu';
@@ -7,11 +9,11 @@ import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles({
   root: {
-    padding: '1rem 4rem 1.5rem 1.5rem',
+    padding: '1rem 4rem 1.5rem 1.5rem'
   },
 });
 
-export default function Design() {
+export default function Category() {
   const classes = useStyles();
   const [anchorEl, setAnchorEl] = React.useState(null);
 
@@ -26,9 +28,10 @@ export default function Design() {
   return (
     <div>
       <Button aria-controls="simple-menu" aria-haspopup="true" onClick={handleClick}>
-        Design Dropdown
+        Category
       </Button>
-      <Menu id="simple-menu" anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleClose}>
+      <Menu id="simple-menu" anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleClose} >
+
         <MenuItem className={classes.root} onClick={handleClose}>
           {' '}
           <Typography variant="body2">All Goals</Typography>
@@ -36,18 +39,16 @@ export default function Design() {
 
         <MenuItem className={classes.root} onClick={handleClose}>
           {' '}
-          <Typography variant="body2">Finance</Typography>
+          <Typography variant="body2">Zuri Workspace</Typography>
         </MenuItem>
 
         <MenuItem className={classes.root} onClick={handleClose}>
-          {' '}
-          <Typography variant="body2">Project Design</Typography>
+          <Typography variant="body2">Private</Typography>
         </MenuItem>
 
-        <MenuItem className={classes.root} onClick={handleClose}>
-          <Typography variant="body2">Wireframing</Typography>
-        </MenuItem>
       </Menu>
     </div>
   );
 }
+
+
