+++ client/src/components/Sortoptions/Progress.js
@@ -1,6 +1,4 @@
-
 import React from 'react';
-
 import { Typography } from '@material-ui/core';
 import Button from '@material-ui/core/Button';
 import Menu from '@material-ui/core/Menu';
@@ -9,11 +7,11 @@ import { makeStyles } from '@material-ui/core/styles';
 
 const useStyles = makeStyles({
   root: {
-    padding: '1rem 4rem 1.5rem 1.5rem'
+    padding: '1rem 4rem 1.5rem 1.5rem',
   },
 });
 
-export default function MobilePrivate() {
+export default function Progress() {
   const classes = useStyles();
   const [anchorEl, setAnchorEl] = React.useState(null);
 
@@ -28,10 +26,9 @@ export default function MobilePrivate() {
   return (
     <div>
       <Button aria-controls="simple-menu" aria-haspopup="true" onClick={handleClick}>
-        Mobile Private Dropdown
+       Progress
       </Button>
-      <Menu id="simple-menu" anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleClose} >
-
+      <Menu id="simple-menu" anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleClose}>
         <MenuItem className={classes.root} onClick={handleClose}>
           {' '}
           <Typography variant="body2">All Goals</Typography>
@@ -39,16 +36,18 @@ export default function MobilePrivate() {
 
         <MenuItem className={classes.root} onClick={handleClose}>
           {' '}
-          <Typography variant="body2">Zuri Workspace</Typography>
+          <Typography variant="body2">Finance</Typography>
         </MenuItem>
 
         <MenuItem className={classes.root} onClick={handleClose}>
-          <Typography variant="body2">Private</Typography>
+          {' '}
+          <Typography variant="body2">Project Design</Typography>
         </MenuItem>
 
+        <MenuItem className={classes.root} onClick={handleClose}>
+          <Typography variant="body2">Wireframing</Typography>
+        </MenuItem>
       </Menu>
     </div>
   );
 }
-
-
